+++ src/main/java/com/wrlimit/planerbackend/repository/CategoryMongoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryMongoRepository extends MongoRepository<Category, Long> {
+public interface CategoryMongoRepository extends MongoRepository<Category, Integer> {
     Category findByName(String name);
 }
