+++ src/main/java/com/wrlimit/planerbackend/repository/TaskMongoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TaskMongoRepository extends MongoRepository<Task, Long> {
+public interface TaskMongoRepository extends MongoRepository<Task, Integer> {
     Task findByName(String name);
 }
